Hypothesis proving
This commit adds a feature named "hypothesis pruning", whereby once a
conjecture has been approved in the present theory, we arbitrarily remove
hypotheses from the conjecture in an attempt to prove and admit a more
"general" form. The expectation is that the search space for pruned
hypotheses is relatively small; however, the naive search used right now
means that if a conjecture that can be proved without any
qualifications acquires 5 hypotheses during conjecture generation, we
will have to search 5!=120 conjectures before concluding that the
theorem is indeed sound without any hypothesis. A trivial way to improve
this performance is to memoize the hypotheses used in the conjecture,
perhaps via a set that records lists of integers, where each integer is
the index of the original hypothesis used.