Prune unsatisfiable conjectures prior to proving, elaborate issues
This commit makes sure that we try to prune conjectures by checking
if they are unsatisfiable in some regard prior to trying to prove them.
This decrease the number of theorems to consider for proof immensily; as
an example,

```
(genM '(CONS LEN BINARY-APPEND REPEAT) 5)
```

dropped from trying to prove 40K theorems to 24K. Obviously there is
still much more to be done to remove the search space, but this is a
pretty good win.

Additionally this commit adds some notes on issues I have noticed with
attempts to prune conjectures due to unsatisfiability of variables with
regards to type hypotheses. The issue is elaborated in detail in
issues/weak-type-set-inference. The TL;DR is that apparently the
type-set inference algorithm is not very strong, relying primarily on
type prescriptions and guards, and does not actually do any proving, so
it will infer certain variables to have an inhabited type set when in
fact no value inhabiting that type set could ever satisfy the
hypothesis. To deal with this, perhaps we can try proving the validity
of hypotheses incrementally (though this is likely to get prohibitively
expensive, at least on a surface guess), or calling out to an SMT solver
via SMTlink.