Get a bit smarter about proving hyps unsatisfiability
This commit introduces a better algorithm for determining if a
hypothesis is never satisfiable. In particular, if A is not satisfiable,
it must be the case that -A is valid. To this end, our theorem-pruning
prepass now also attempt to prove the validity of (not hypotheses), and
it this is valid, the theorem is discarded as "useless". We still keep
the type-satisfiability check there too since it is quite fast and can
serve as a "fast path".

Unfortunately this is still not enough for discarding things that we may
believe are pretty useless. For example consider the hypothesis

```
(and (listp b) (natp (append b a)))
```

Certainly this is not satisfiable by any `b`, but the prover has trouble
showing that the converse holds.