Extract hypotheses from existing theory database
This commit attempts to improve the quality and amount of theorems we
are able to generate by attempting to extract reasonable-looking
hypotheses from the present world. The idea is to look at existing
lemmas in the database, and see what hypotheses terms in the LHS of
those lemmas use. This information is stored in a table of
term->hypotheses->[extra data]. Then, when we have a conjecture that
perhaps fails to be proven in a first-pass with hypotheses introduced
from the somewhat-naive "Hypothesis Collection" phase (which is mostly
just adding hyps from function guards), we can try to add these
"extracted hypotheses" to the conjecture, if there are any matching ones
at all.

Preliminary experimentation suggests this actually might not be all that
useful, unfortunately. At least for a small theory, the hypotheses we
are able to extract seem to be few and far between. However, I am not
sure what else we can do to introduce hypotheses at this point, so
let's just give it a shot :shrug:.